refactor/rename Publisher -> Dispenser, add Logger
SMS Publisher has been renamed to SMS Dispenser in order to avoid
confusion with PUB/SUB (publisher/subscriber) sockets, which will be
added soon (SMS Acceptor already has an appropriate name (it's not
called Subscriber) so no change is needed there).

In the process, SMS Dispenser has been thoroughly refactored so that
all common "dispensing" code is in sms_dispenser.rb while all the
provider-specific code is in sms_dispenser-plivo.rb (since Plivo is
the only provider that has been ported so far).  The next step is to
port sms_relay-pubsub.rb to sms_dispenser-vitelity.rb so we have both
providers migrated to the new codebase.  This will set us up nicely
for the migration to PUB/SUB sockets, which in turn will make it much
easier to fix the hack we made in 09b8504 to add a Vitelity frontend.

A new Logger module has been added, which includes the log and log_raw
methods that previously (and frighteningly) were copied into each new
module.  For now the Logger is only used by SMS Dispenser, but it will
soon be replacing all other logging code.