add test infrastructure and an initial test case
To integrate with this new handmade test infrastructure, the Relay
applications needed to be updated to accept a host and port (so we
could use a different one than the Jabber ID defined, which is needed
when connecting to the XMPP server running on localhost).

The rest of this commit consists of the test infrastructure and a
single test case (messaging-send_sms-simple.sh).  Additional test
cases will be added shortly.  See the updated README to run the tests.

To add a new test, create an executable file in the test directory and
add its filename to the test_list file.  The test should provide an
appropriate return code to indicate whether the test passed or failed.